Remove reference to IN_PRODUCTION env var
This was just me being stupid. If running locally
using the built-in server, the app will automatically
run in debug mode because the __name__ == '__main__'
suite will be run. If we are on a production server,
this suite won't be run (as the application is running
inside a WSGI container) and so `debug` will be set to its
default value - which is False. The environment variable
approach is only needed if we need to use different
config files in different environments, which we don't.